# src/main/java/de/getdate/backend/dto/AddEmployeeRequest.java
package de.getdate.backend.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class AddEmployeeRequest {
  private String name;

}