# src/main/java/de/getdate/backend/dto/RegisterRequest.java
package de.getdate.backend.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class RegisterRequest {
  private String email;
  private String firstname;
  private String lastname;
  private String password;

}